+++ app/page.tsx
@@ -1,7 +1,7 @@
-'use client'
+"use client";
 import React from "react";
 import { Card, CardBody, CardFooter, Image } from "@nextui-org/react";
-import Imagen from "@/components/types/image/image"; // para zoom pendiente por implementar 
+import Imagen from "@/components/types/svg/image"; // para zoom pendiente por implementar
 
 export default function App() {
   const list = [
@@ -45,14 +45,97 @@ export default function App() {
       img: "https://nextui.org/images/fruit-8.jpeg",
       price: "$12.20",
     },
+    {
+      title: "Orange",
+      img: "https://nextui.org/images/fruit-1.jpeg",
+      price: "$5.50",
+    },
+    {
+      title: "Tangerine",
+      img: "https://nextui.org/images/fruit-2.jpeg",
+      price: "$3.00",
+    },
+    {
+      title: "Raspberry",
+      img: "https://nextui.org/images/fruit-3.jpeg",
+      price: "$10.00",
+    },
+    {
+      title: "Lemon",
+      img: "https://nextui.org/images/fruit-4.jpeg",
+      price: "$5.30",
+    },
+    {
+      title: "Avocado",
+      img: "https://nextui.org/images/fruit-5.jpeg",
+      price: "$15.70",
+    },
+    {
+      title: "Lemon 2",
+      img: "https://nextui.org/images/fruit-6.jpeg",
+      price: "$8.00",
+    },
+    {
+      title: "Banana",
+      img: "https://nextui.org/images/fruit-7.jpeg",
+      price: "$7.50",
+    },
+    {
+      title: "Watermelon",
+      img: "https://nextui.org/images/fruit-8.jpeg",
+      price: "$12.20",
+    },
+    {
+      title: "Orange",
+      img: "https://nextui.org/images/fruit-1.jpeg",
+      price: "$5.50",
+    },
+    {
+      title: "Tangerine",
+      img: "https://nextui.org/images/fruit-2.jpeg",
+      price: "$3.00",
+    },
+    {
+      title: "Raspberry",
+      img: "https://nextui.org/images/fruit-3.jpeg",
+      price: "$10.00",
+    },
+    {
+      title: "Lemon",
+      img: "https://nextui.org/images/fruit-4.jpeg",
+      price: "$5.30",
+    },
+    {
+      title: "Avocado",
+      img: "https://nextui.org/images/fruit-5.jpeg",
+      price: "$15.70",
+    },
+    {
+      title: "Lemon 2",
+      img: "https://nextui.org/images/fruit-6.jpeg",
+      price: "$8.00",
+    },
+    {
+      title: "Banana",
+      img: "https://nextui.org/images/fruit-7.jpeg",
+      price: "$7.50",
+    },
+    {
+      title: "Watermelon",
+      img: "https://nextui.org/images/fruit-8.jpeg",
+      price: "$12.20",
+    },
   ];
 
   return (
-
-
-    <div className="gap-2 grid grid-cols-2 sm:grid-cols-4">
+    <div className="gap-2 grid grid-cols-2 sm:grid-cols-3">
       {list.map((item, index) => (
-        <Card shadow="sm" key={index} isPressable onPress={() => console.log("item pressed")}>
+        <Card
+          shadow="sm"
+          key={index}
+          isPressable
+          onPress={() => console.log("item pressed")}
+        >
           <CardBody className="overflow-visible p-0">
             <Image
               shadow="sm"
