+++ components/types/svg/image.tsx
