+++ components/types/svg/logo.tsx
